refactor: extract route delay info into separate view
